refactor(transfer): Architect for lazy loading and server metadata
This refactor resolves the architectural conflict between server-rendered `metadata` and the need for client-side lazy loading with `ssr: false`.

By introducing dedicated "bridge" components (`LazySendPage`, `LazyReceivePage`), this change allows the parent `page.tsx` to remain a Server Component, responsible solely for exporting SEO-critical `metadata`. The heavy, interactive UI (`Send.tsx`, `Receive.tsx`) is now correctly deferred and rendered on the client, significantly improving initial page load performance.

Key Changes:
- Parent pages (`page.tsx`) now handle only server-side `metadata`.
- New client "bridge" components (`Lazy...`) manage the `dynamic()` import with `ssr: false`.
- Core UI components are now loaded on-demand on the client, showing a loading state to improve perceived performance.

Ultimately, this structure allows the application to benefit from both strong SEO and superior initial load performance without compromise.
