+++ relayr-ui/src/app/transfer/receive/LazyReceivePage.tsx
@@ -1,0 +1,17 @@
+"use client";
+
+// Next.js
+import dynamic from "next/dynamic";
+
+// Custom Components
+import ExperienceLoading from "@/components/loading/ExperienceLoading";
+
+// Lazy-loaded Internal Components
+const Receive = dynamic(() => import("./Receive"), {
+  loading: () => <ExperienceLoading />,
+  ssr: false,
+});
+
+export default function LazyReceivePage() {
+  return <Receive />;
+}
